+++ Tactical/LootEverything/mod_info.js
@@ -3,7 +3,7 @@
    Lang : "-",
    Duration : "temp",
    Name : "Loot Everything",
-   Description : "Make all enemies drop all loots.",
+   Description : "Make all enemies drop all loots, including armour by default.\n\nKnown side effect: soldiers will be stripped of armour on death.",
    Url : {
       "Nexus" : "https://nexusmods.com/phoenixpoint/mods/33/",
       "GitHub" : "https://github.com/Sheep-y/PhoenixPt-Mods/",