+++ Tactical/LootEverything/changelog.md
@@ -1,6 +1,7 @@
 Changelog of Loot Everything, a Phoenix Point Mod by Sheepy
 
-# Version 1, 2020-06-30
+# Version 1, 2020-07-01
 
 * Force-drop all droppable equipments on die.
+* Force-drop all armours and mounted items on die.
 * Tried doing configurable mod such as bonus drop chance and drop chance multiplier, end up decided to keep it simple, bullet proof and fail fast.